Allow overriding the command line when pinning to the taskbar
Windows 7 and later allow applications to be pinned to the taskbar. In
particular when calling mintty as the work horse of the Git Bash, we
will want to re-launch the Git Bash when being pinned, not just mintty.

Let's just add a setting to allow users to override which command gets
to be launched when the pinned mintty is called.

Note: this should be used in conjunction with the AppID setting.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>